source cleanups [wip]
