--- .github/workflows/run-tests.yml
@@ -1,54 +1,0 @@
-name: Run tests
-
-on:
-  push:
-    branches:
-    - main
-    paths:
-    - '**.swift'
-  pull_request:
-    branches:
-    - main
-
-jobs:
-  
-  macOS-tests:
-    runs-on: self-hosted
-    steps:
-    
-    - name: Checkout
-      uses: actions/checkout@v4
-    
-    # - name: Cache
-    #   uses: actions/cache@v3
-    #   with:
-    #     path: server/.build
-    #     key: ${{ runner.os }}-spm-${{ hashFiles('**/Package.resolved') }}
-    #     restore-keys: ${{ runner.os }}-spm-
-    
-    - name: Test
-      run: swift test --parallel --enable-code-coverage
-  
-  linux-tests:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        image:
-          - 'swift:5.9'
-          - 'swift:5.10'
-    container:
-      image: ${{ matrix.image }}
-    steps:
-    
-    - name: Checkout
-      uses: actions/checkout@v4
-    
-    # - name: Cache
-    #   uses: actions/cache@v3
-    #   with:
-    #     path: server/.build
-    #     key: ${{ runner.os }}-spm-${{ hashFiles('**/Package.resolved') }}
-    #     restore-keys: ${{ runner.os }}-spm-
-    
-    - name: Test
-      run: swift test --parallel --enable-code-coverage
