+++ .github/workflows/testing.yml
@@ -1,0 +1,39 @@
+name: Testing
+
+on:
+  pull_request:
+    branches:
+    - main
+
+jobs:
+
+  swiftlang_checks:
+    name: Swiftlang Checks
+    uses: swiftlang/github-workflows/.github/workflows/soundness.yml@main
+    with:
+      license_header_check_project_name: "project"
+      format_check_enabled : true
+      broken_symlink_check_enabled : true
+      unacceptable_language_check_enabled : true
+      shell_check_enabled : true
+      docs_check_enabled : false
+      api_breakage_check_enabled : false
+      license_header_check_enabled : false
+      yamllint_check_enabled : false
+      python_lint_check_enabled : false
+
+  bb_checks:
+    name: BB Checks
+    uses: BinaryBirds/github-workflows/.github/workflows/extra_soundness.yml@main
+    with:
+      local_swift_dependencies_check_enabled : true
+      headers_check_enabled : true
+      docc_warnings_check_enabled : true
+
+  swiftlang_tests:
+    name: Swiftlang Tests
+    uses: swiftlang/github-workflows/.github/workflows/swift_package_test.yml@main
+    with:
+      enable_windows_checks : false
+      linux_build_command: "swift test --parallel --enable-code-coverage"
+      linux_exclude_swift_versions: "[{\"swift_version\": \"5.8\"}, {\"swift_version\": \"5.9\"}, {\"swift_version\": \"5.10\"}, {\"swift_version\": \"nightly\"}, {\"swift_version\": \"nightly-main\"}, {\"swift_version\": \"6.0\"}, {\"swift_version\": \"nightly-6.0\"}, {\"swift_version\": \"nightly-6.1\"}, {\"swift_version\": \"nightly-6.3\"}]"
