+++ .swiftformatignore
@@ -1,0 +1,1 @@
+Package.swift
