+++ LICENSE
@@ -1,7 +1,7 @@
 MIT License
 
 Copyright (c) 2018-2022 Tibor Bödecs
-Copyright (c) 2022-2024 Binary Birds Ltd.
+Copyright (c) 2022-2026 Binary Birds Kft.
 
 Permission is hereby granted, free of charge, to any person
 obtaining a copy of this software and associated documentation
