+++ Makefile
@@ -1,22 +1,41 @@
 SHELL=/bin/bash
 
-build:
-	swift build
+baseUrl = https://raw.githubusercontent.com/BinaryBirds/github-workflows/refs/heads/main/scripts
 
-release:
-	swift build -c release
+check: symlinks language deps lint headers
+
+symlinks:
+	curl -s $(baseUrl)/check-broken-symlinks.sh | bash
+
+language:
+	curl -s $(baseUrl)/check-unacceptable-language.sh | bash
+
+deps:
+	curl -s $(baseUrl)/check-local-swift-dependencies.sh | bash
+
+lint:
+	curl -s $(baseUrl)/run-swift-format.sh | bash
+
+format:
+	curl -s $(baseUrl)/run-swift-format.sh | bash -s -- --fix
+
+docc-local:
+	curl -s $(baseUrl)/generate-docc.sh | bash -s -- --local
+
+run-docc:
+	curl -s $(baseUrl)/run-docc-docker.sh | bash
+
+docc-warnings:
+	curl -s $(baseUrl)/check-docc-warnings.sh | bash
+
+headers:
+	curl -s $(baseUrl)/check-swift-headers.sh | bash
+
+fix-headers:
+	curl -s $(baseUrl)/check-swift-headers.sh | bash -s -- --fix
 	
 test:
 	swift test --parallel
 
-test-with-coverage:
-	swift test --parallel --enable-code-coverage
-
-clean:
-	rm -rf .build
-
-check:
-	./scripts/run-checks.sh
-
-format:
-	./scripts/run-swift-format.sh --fix
+docker-test:
+	docker build -t feather-storage-tests . -f ./docker/tests/Dockerfile && docker run --rm feather-storage-tests
