+++ Sources/FeatherStorage/StorageComponent.swift
@@ -5,116 +5,10 @@
 //  Created by Tibor Bodecs on 2020. 04. 28..
 //
 
-import FeatherComponent
 import NIOCore
 
-/// storage chunks returned used by the multipart request apis
-public struct StorageChunk: Hashable, Codable, Sendable, Equatable {
-    public let chunkId: String
-    public let number: Int
-
-    public init(
-        chunkId: String,
-        number: Int
-    ) {
-        self.chunkId = chunkId
-        self.number = number
-    }
-}
-
-/// storage any async sequence
-public struct StorageAnyAsyncSequence<Element>: Sendable, AsyncSequence {
-    public typealias ByteChunk = ArraySlice<UInt8>
-
-    public typealias AsyncIteratorNextCallback = () async throws -> Element?
-    public let length: UInt64?
-
-    public struct AsyncIterator: AsyncIteratorProtocol {
-        @usableFromInline let nextCallback: AsyncIteratorNextCallback
-
-        @inlinable init(nextCallback: @escaping AsyncIteratorNextCallback) {
-            self.nextCallback = nextCallback
-        }
-
-        @inlinable public mutating func next() async throws -> Element? {
-            try await self.nextCallback()
-        }
-    }
-
-    @usableFromInline var makeAsyncIteratorCallback:
-        @Sendable () -> AsyncIteratorNextCallback
-
-    @inlinable public init<SequenceOfBytes>(
-        asyncSequence: SequenceOfBytes,
-        length: UInt64?
-    )
-    where
-        SequenceOfBytes: AsyncSequence & Sendable,
-        SequenceOfBytes.Element == ByteBuffer,
-        Element == ByteChunk
-    {
-        self.makeAsyncIteratorCallback = {
-            var iterator = asyncSequence.makeAsyncIterator()
-            return {
-                if var buffer = try await iterator.next() {
-                    return ArraySlice(
-                        buffer.readBytes(length: buffer.readableBytes) ?? []
-                    )
-                }
-                return nil
-            }
-        }
-        self.length = length
-    }
-
-    @inlinable public init<SequenceOfBytes>(
-        asyncSequence: SequenceOfBytes,
-        length: UInt64?
-    )
-    where
-        SequenceOfBytes: AsyncSequence & Sendable,
-        SequenceOfBytes.Element == ByteChunk,
-        Element == ByteBuffer
-    {
-        self.makeAsyncIteratorCallback = {
-            var iterator = asyncSequence.makeAsyncIterator()
-            return {
-                if let arraySlice = try await iterator.next() {
-                    var byteBuffer = ByteBufferAllocator()
-                        .buffer(capacity: arraySlice.count)
-                    byteBuffer.writeBytes(arraySlice)
-                    return byteBuffer
-                }
-                return nil
-            }
-        }
-        self.length = length
-    }
-
-    @inlinable public init<SequenceOfBytes>(
-        asyncSequence: SequenceOfBytes,
-        length: UInt64?
-    )
-    where
-        SequenceOfBytes: AsyncSequence & Sendable,
-        SequenceOfBytes.Element == Element
-    {
-        self.makeAsyncIteratorCallback = {
-            var iterator = asyncSequence.makeAsyncIterator()
-            return {
-                try await iterator.next()
-            }
-        }
-        self.length = length
-    }
-
-    @inlinable public func makeAsyncIterator() -> AsyncIterator {
-        .init(nextCallback: self.makeAsyncIteratorCallback())
-    }
-}
-
 /// storage component protocol
-public protocol StorageComponent: Component {
+public protocol StorageComponent {
 
     /// returns the available storage space
     var availableSpace: UInt64 { get }
@@ -217,96 +111,4 @@ public protocol StorageComponent: Component {
     ) async throws
 }
 
-public struct StorageByteBufferAsyncSequenceWrapper: Sendable, AsyncSequence {
-    public typealias Element = ByteBuffer
-    let buffer: ByteBuffer
 
-    public init(buffer: ByteBuffer) {
-        self.buffer = buffer
-    }
-
-    public struct AsyncIterator: AsyncIteratorProtocol {
-        let buffer: ByteBuffer
-        var currentIndex: Int = 0
-
-        public mutating func next() async -> ByteBuffer? {
-            guard currentIndex < buffer.readableBytes
-            else {
-                return nil
-            }
-
-            let endIndex = Swift.min(
-                currentIndex + 32 * 1024,
-                buffer.readableBytes
-            )
-            let chunkRange = currentIndex..<endIndex
-
-            var chunk = buffer
-            chunk.moveReaderIndex(to: chunkRange.lowerBound)
-            chunk.moveWriterIndex(to: chunkRange.upperBound)
-
-            currentIndex = endIndex
-            return chunk
-        }
-    }
-
-    public func makeAsyncIterator() -> AsyncIterator {
-        AsyncIterator(buffer: buffer)
-    }
-}
-
-extension StorageComponent {
-
-    public func move(
-        key source: String,
-        to destination: String
-    ) async throws {
-        let exists = await exists(key: source)
-        guard exists else {
-            throw StorageComponentError.invalidKey
-        }
-        try await copy(key: source, to: destination)
-        try await delete(key: source)
-    }
-
-    public func upload(
-        key: String,
-        buffer: ByteBuffer
-    ) async throws {
-        try await uploadStream(
-            key: key,
-            sequence: .init(
-                asyncSequence: StorageByteBufferAsyncSequenceWrapper(
-                    buffer: buffer
-                ),
-                length: UInt64(buffer.readableBytes)
-            )
-        )
-    }
-
-    public func download(
-        key: String,
-        range: ClosedRange<Int>?
-    ) async throws -> ByteBuffer {
-        try await downloadStream(key: key, range: range).collect(upTo: Int.max)
-    }
-
-    public func upload(
-        multipartId: String,
-        key: String,
-        number: Int,
-        buffer: ByteBuffer
-    ) async throws -> StorageChunk {
-        try await uploadStream(
-            multipartId: multipartId,
-            key: key,
-            number: number,
-            sequence: .init(
-                asyncSequence: StorageByteBufferAsyncSequenceWrapper(
-                    buffer: buffer
-                ),
-                length: UInt64(buffer.readableBytes)
-            )
-        )
-    }
-}
